Add functions to compute extremas on edge
So far only the extrema in the middle of our data is used for the
bspline interpolation. This computes the edge extrema as suggested by Wu
and Huang. The extrema at the edge is computed as either the data at the
edge or as the linear interpolation of the last two extrema whatever is
more extreme.