Add function get_edgepoint
The function get_edgepoint includes the functionality of the functions
startmin, startmax, endmin, endmax, it will give a point on the edge of
the time series on either sites which is interpolated from the last two
extrema or the point on the edge. This stabelizes the edgecase further.
I'm not sure if the branching on the first and last function is a good
idea but I want to try it at the moment.